+++ Coimbra.Editor/SceneProcessingSettings.cs
@@ -7,7 +7,7 @@ namespace Coimbra.Editor
     /// <summary>
     /// Settings to customize the behaviour of <see cref="ISceneProcessorComponent"/> and <see cref="IScenePostProcessorComponent"/>.
     /// </summary>
-    [ProjectSettings(CoimbraUtility.ProjectSettingsPath, true)]
+    [ProjectSettings(CoimbraUtility.ProjectSettingsPath, true, FileDirectory = CoimbraUtility.ProjectSettingsFilePath)]
     public sealed class SceneProcessingSettings : ScriptableSettings
     {
         /// <inheritdoc cref="IOrderedCallback.callbackOrder"/>